requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# GPU & Deep Learning
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Image Processing
pillow>=10.0.0

# File Upload Support
python-multipart>=0.0.6

# HTTP Client (for testing)
httpx>=0.25.0

# Additional utilities
numpy>=1.24.0
